Perform automatic migrations in development
When database schemas change in development environments, it is
irritating to have to manually perform a database migration or restart
the server to have it apply the necessary migrations. The new database
access library uses an internal wrapper function that detects failures
because a database migration is necessary and runs available
migrations. While migrations still need to be generated manually, they
will be applied automatically when rendering the first page that needs
them.