Fix linting errors on ~ imports
Remix by convention seems to use ~ as a prefix to indicate the project
root for import paths. This is configured in the tsconfig.json, which
is why code built successfully, but ESLint was throwing warnings
because it was not respecting that part of the tsconfig.json. Adding
the eslint-import-resolver-typescript package as a development package
lets the ESLint TypeScript plugin use the paths configuration from
tsconfig.json and fixes the errors when linting.