Downgrade @remix-run/serve to 2.3.1
When trying to access the Zod errors, it somehow trips the installed
sourcemapping which fails and throws an error because Zod doesn't ship
with source maps. This bug (or at least a very similar one) is being
tracked in https://github.com/remix-run/remix/issues/8309, but until it
is resolved, I'm pinning the version of the Remix server to 2.3.1 since
that is the latest version that is not affected.